feat: add 2027 ski trip mountain info
To make this reusable in the future, a page is specifically dedicated
for the '26-'27 ski season. This will contain all data related to the
mountains we will choose from to ski.

Some basic schema validation is necessary and a base for the mountain
data is created here. We want to know some basic data about the
mountains as well as estimated costs for various aspects of the ski trip.