fix: change skew of cloud text and fix gap
Instead of wrapping behind the mountain, the title and subtitle should
display in front of the mountain and wrap around it. This means some
tweaking of the position of the container div is required as the
transform origins have changed, which had caused the text to appear off
center.

There was also a gap on xl screens between the ski slopes and the
mountain peak which has been fixed by increasing the y translation for
this screen size.